+++ internal/provider/data_source_fmc_access_control_policy.go
@@ -305,24 +305,12 @@ func (d *AccessControlPolicyDataSource) Schema(ctx context.Context, req datasour
 							Computed:            true,
 							NestedObject: schema.NestedAttributeObject{
 								Attributes: map[string]schema.Attribute{
-									"type": schema.StringAttribute{
-										MarkdownDescription: "Type of the object.",
-										Computed:            true,
-									},
-									"port": schema.StringAttribute{
-										MarkdownDescription: "Port number.",
-										Computed:            true,
-									},
 									"protocol": schema.StringAttribute{
 										MarkdownDescription: "IANA protocol number.",
 										Computed:            true,
 									},
-									"icmp_type": schema.StringAttribute{
-										MarkdownDescription: "ICMP type.",
-										Computed:            true,
-									},
-									"icmp_code": schema.StringAttribute{
-										MarkdownDescription: "ICMP code.",
+									"port": schema.StringAttribute{
+										MarkdownDescription: "Port number.",
 										Computed:            true,
 									},
 								},