+++ internal/provider/data_source_fmc_access_rule.go
@@ -209,24 +209,12 @@ func (d *AccessRuleDataSource) Schema(ctx context.Context, req datasource.Schema
 				Computed:            true,
 				NestedObject: schema.NestedAttributeObject{
 					Attributes: map[string]schema.Attribute{
-						"type": schema.StringAttribute{
-							MarkdownDescription: "Type of the object.",
-							Computed:            true,
-						},
-						"port": schema.StringAttribute{
-							MarkdownDescription: "Port number.",
-							Computed:            true,
-						},
 						"protocol": schema.StringAttribute{
 							MarkdownDescription: "IANA protocol number.",
 							Computed:            true,
 						},
-						"icmp_type": schema.StringAttribute{
-							MarkdownDescription: "ICMP type.",
-							Computed:            true,
-						},
-						"icmp_code": schema.StringAttribute{
-							MarkdownDescription: "ICMP code.",
+						"port": schema.StringAttribute{
+							MarkdownDescription: "Port number.",
 							Computed:            true,
 						},
 					},